# gh-117431: Improve performance of startswith, endswith, count and find methods for strings and bytes