--- a/Misc/NEWS.d/next/Core and Builtins/2024-04-01-21-15-35.gh-issue-117431.UBT4zj.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2024-04-01-21-15-35.gh-issue-117431.UBT4zj.rst
@@ -1,0 +1,1 @@
+Improve performance of startswith, endswith, count and find methods for strings and bytes.