Benchmarks (non-PGO):

```
main: python -m timeit -s "s = 'abcdef'" "s.startswith('abc')"
5000000 loops, best of 5: 63.8 nsec per loop

pr: python -m timeit -s "s = 'abcdef'" "s.startswith('abc')"
5000000 loops, best of 5: 42.3 nsec per loop
```

```
main: python -m timeit -s "s = 'abcdef'" "s.count('a')"
5000000 loops, best of 5: 70 nsec per loop

pr: python -m timeit -s "s = 'abcdef'" "s.count('a')"
5000000 loops, best of 5: 51.4 nsec per loop
```

```
main: python -m timeit -s "s = b'abcdef'" "s.startswith(b'q')"
5000000 loops, best of 5: 68.3 nsec per loop

pr: python -m timeit -s "s = b'abcdef'" "s.startswith(b'q')"
5000000 loops, best of 5: 44.5 nsec per loop
```

Notes:
* We could also add a fast path for the case where there are 2 or 3 arguments. Since all arguments are positional only, for 1, 2 or 3 arguments the parsing cannot fail. That will increase performance for the 2 and 3 argument case, but it adds some more complexity.
 
<!--
Thanks for your contribution!
Please read this comment in its entirety. It's quite important.

# Pull Request title

It should be in the following format:

```
gh-NNNNN: Summary of the changes made
```

Where: gh-NNNNN refers to the GitHub issue number.

Most PRs will require an issue number. Trivial changes, like fixing a typo, do not need an issue.

# Backport Pull Request title

If this is a backport PR (PR made against branches other than `main`),
please ensure that the PR title is in the following format:

```
[X.Y] <title from the original PR> (GH-NNNN)
```

Where: [X.Y] is the branch name, e.g. [3.6].

GH-NNNN refers to the PR number from `main`.

-->


<!-- gh-issue-number: gh-117431 -->
* Issue: gh-117431
<!-- /gh-issue-number -->